Safe Test Runner for Mongoose Models (#1070)
* added setup for safe tests'

* added error for serverCleanup

* refactoring setupSafeTest

* added tests

* changed test

* added test consts

* Refactor: moved envs to centralized config for consistency

* Update src/test/setupSafeTest.js


---------

+++ src/test/setup.js
@@ -2,6 +2,9 @@ const express = require('express')
 const mongoose = require('mongoose')
 const request = require('supertest')
 require('~/initialization/envSetup')
+const { createError } = require('~/utils/errorsHelper')
+const { RESTRICTED_PATH } = require('~/consts/errors')
+const { isRestrictedPath } = require('~/utils/isRestrictedPath')
 
 const serverSetup = require('~/initialization/serverSetup')
 
@@ -12,6 +15,10 @@ const serverInit = async () => {
 }
 
 const serverCleanup = async () => {
+  if (isRestrictedPath('/models/')) {
+    throw createError(403, RESTRICTED_PATH('serverCleanup', '/models/'))
+  }
+
   await mongoose.connection.db.dropDatabase()
 }
